# src/main/resources/api/api.yml
openapi: 3.1.0
info:
  title: Football championship API
  description: Football
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
...
  /clubs:
    get:
      tags:
        - Club
      summary: Get clubs of the championship
      operationId: getClubs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            description: page number
          required: false
        - name: size
          in: query
          schema:
            type: integer
            default: 10
            description: element for each page
          required: false
      responses:
        200:
          description: List of the clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
    put:
      tags:
        - Club
      summary: Create new clubs or update if already exist
      operationId: createOrUpdateClubs
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Club'
      responses:
        200:
          description: List of created or updated clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
...
components:
...
  schemas:
    Club:
      allOf:
        - $ref: '#/components/schemas/ClubMinimumInfo'
      type: object
      properties:
        yearCreation:
          type: integer
        stadium:
          type: string
        coach:
          $ref: '#/components/schemas/Coach'
    Coach:
      type: object
      properties:
        name:
          type: string
        nationality:
          type: string
...
    ClubMinimumInfo:
      type: object
      properties:
        id:
          type: string
          description: Club identifier
        name:
          type: string
        acronym:
          type: string
          description: Three letters describing the club (not-unique)
          example: RMA
...